[CHORE] 프로젝트 전체조회 API Response 추가(member, 창업중, 진행중) (#318)
* refactor: booleanexpression 메소드 형태로 변경

* fix: count method by filter

* chore: refactor method

* chore: response type

* add: 창업중, 진행중 response 추가

* chore: member response 추가

+++ src/main/java/org/sopt/makers/internal/dto/project/ProjectAllResponse.java
@@ -5,5 +5,5 @@ import java.util.List;
 public record ProjectAllResponse(
         List<ProjectResponse> projectList,
         Boolean hasNext,
-        Long totalCount
+        Integer totalCount
 ){}
